SCRUM-118 refactor: file structure refactoring
