--- src/entities/posts/ui/EmblaCarousel/EmblaCarousel.module.scss
@@ -1,96 +1,0 @@
-.embla {
-  position: relative;
-  overflow: hidden;
-  border-radius: 2px;
-  //margin-bottom: 1rem;
-}
-
-.embla__viewport {
-  overflow: hidden;
-  width: 100%;
-}
-
-.embla__container {
-  display: flex;
-}
-
-.embla__slide {
-  flex: 0 0 100%;
-  min-width: 0;
-}
-
-.embla__slide img {
-  width: 100%;
-  height: auto;
-  //max-height: 400px;
-  object-fit: cover;
-  border-bottom: 1px solid #222;
-}
-
-.embla__button {
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  background: rgba(50, 50, 50, 0.6);
-  color: #fff;
-  font-size: 32px;
-  border: none;
-  cursor: pointer;
-  padding: 8px 14px;
-  border-radius: 4px;
-  z-index: 2;
-  transition: background 0.2s;
-}
-
-.embla__button:hover {
-  background: rgba(30, 30, 30, 0.9);
-}
-
-.embla__button__prev {
-  left: 10px;
-}
-
-.embla__button__next {
-  right: 10px;
-}
-
-.embla__dots {
-  position: absolute;
-  bottom: 12px;
-  left: 50%;
-  transform: translateX(-50%);
-  background: #333333;
-  opacity: 50%;
-  padding: 4px 10px;
-  border-radius: 2px;
-  display: flex;
-  gap: 6px;
-  cursor: pointer;
-}
-
-.embla__dot {
-  width: 8px;
-  height: 8px;
-  border-radius: 50%;
-  background: #bbb;
-}
-
-.embla__dot__active {
-  background: #1976d2;
-}
-
-.clarendon {
-  filter: contrast(120%) saturate(125%);
-}
-
-.lark {
-  filter: brightness(110%) saturate(105%);
-}
-
-.gingham {
-  filter: contrast(90%) brightness(105%) sepia(10%);
-}
-
-.moon {
-  filter: grayscale(100%) contrast(110%);
-}
