+++ src/entities/posts/ui/PostModal/DeletePostModal/DeletePostModal.module.scss
