+++ src/entities/posts/ui/PostModal/DeletePostModal/DeletePostModal.tsx
