+++ src/entities/posts/ui/PostModal/EditMode/ExitConfirmationModal/ExitConfirmationModal.tsx
