+++ src/entities/posts/ui/PostModal/PostModal.module.scss
