+++ src/entities/posts/ui/PostModal/ViewMode/ViewMode.module.scss
