+++ src/entities/posts/ui/PostModal/ViewMode/ViewMode.tsx
