+++ src/entities/posts/ui/PostModal/ViewMode/ViewModePhotoSection/ViewModePhotoSection.tsx
