+++ src/entities/posts/ui/PostModal/ViewMode/ViewModePostFooter/ViewModePostFooter.tsx
