+++ src/entities/posts/ui/PostModal/ViewMode/ViewModePostHeader/ViewModePostHeader.tsx
